mpi/pmix: PMIx_Abort avoid keeping abort status in `pmixp_info`
Since the structure `pmixp_info` is specific only to stepdemons,
it makes sense to keep the abort status separately (only for `srun`).
In additional, the init procedure of `pmixp_info` does not perform
for `srun` context, which requires to init `abort_status` out the
side thereby breaking a `pmixp_info` init concept.

Signed-off-by: Boris Karasev <karasev.b@gmail.com>